backend/requirements.txt

uvicorn[standard]==0.29.0
sqlalchemy>=2.0.35

python-jose[cryptography]==3.3.0
bcrypt==4.1.3
python-multipart==0.0.9
scikit-learn>=1.5.0
numpy>=2.0.0
pandas==2.2.2
fastapi>=0.115.0
pydantic>=2.10.0
